requirements.txt
streamlit
openai
langchain
langchain-core
langchain-openai
langchain-community
faiss-cpu
pdfplumber
pytesseract
Pillow
pdf2image
camelot-py[cv]
python-dotenv
tqdm
pandas
numpy
scikit-learn
rank_bm25
sentence-transformers
plotly
